requirements.txt
environ-config~=20.1
aioxmpp~=0.12
